feat: show modal and move page after creating challenge
